+++ src/buffer.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 #include <stdbool.h>
 
-bool adjust_recv_buffer(char **buf, int *buf_size, size_t data_length);
-bool adjust_send_buffer(char **buf, int *buf_size, size_t required_size);
+bool adjust_recv_buffer(char **buf, size_t *buf_size, size_t data_length);
+bool adjust_send_buffer(char **buf, size_t *buf_size, size_t required_size);
 
 #endif
